requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
pillow>=9.0.0
scipy>=1.10.0
matplotlib>=3.5.0
tqdm>=4.65.0
numpy>=1.23.0

# Optional dependencies
tensorboard>=2.13.0
opencv-python>=4.7.0
